chore(deps): upgrade Rollup to latest version
