+++ package.json
@@ -29,6 +29,10 @@
     "@babel/preset-react": "^7.24.1",
     "@babel/preset-typescript": "^7.24.1",
     "@octokit/rest": "^20.0.0",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-terser": "^0.4.4",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "14.5.2",
@@ -58,6 +62,9 @@
     "react": "^18",
     "react-dom": "^18",
     "rimraf": "^5.0.5",
+    "rollup": "^4.17.2",
+    "rollup-plugin-scss": "^4.0.0",
+    "sass": "^1.77.2",
     "surge": "^0.23.1",
     "ts-node": "^10.9.1",
     "ts-patch": "^2.1.0",
