+++ packages/react-core/.npmignore
@@ -1,3 +1,3 @@
 dist/*.tsbuildinfo
 tsconfig.*
-rollup.config.js
+rollup.config.mjs
