+++ packages/react-core/package.json
@@ -55,16 +55,9 @@
   },
   "devDependencies": {
     "@patternfly/patternfly": "5.4.0-prerelease.3",
-    "@rollup/plugin-commonjs": "^25.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.2",
-    "@rollup/plugin-replace": "^5.0.2",
     "css": "^2.2.3",
     "fs-extra": "^11.1.1",
-    "glob": "^7.1.2",
-    "rollup": "^3.21.5",
-    "rollup-plugin-scss": "^4.0.0",
-    "rollup-plugin-svg": "2.0.0",
-    "rollup-plugin-terser": "^7.0.2"
+    "glob": "^7.1.2"
   },
   "peerDependencies": {
     "react": "^17 || ^18",
