--- packages/react-core/rollup.config.js
@@ -1,7 +1,0 @@
-const { name } = require('./package.json');
-const baseConfig = require('../rollup.base');
-
-module.exports = baseConfig({
-  packageName: name.replace('@patternfly/', ''),
-  name: 'PatternFlyReact'
-});
